chore: UI edits based on fri meeting
