+++ src/components/VoucherScanner/ManualAddVoucherModal.tsx
@@ -1,43 +1,20 @@
 import React, { FunctionComponent, useState } from "react";
-import {
-  StyleSheet,
-  View,
-  Modal,
-  TouchableWithoutFeedback,
-  TouchableOpacity
-} from "react-native";
+import { StyleSheet, View } from "react-native";
 import { size, color } from "../../common/styles";
 import { InputWithLabel } from "../Layout/InputWithLabel";
-import { Card } from "../Layout/Card";
 import { DarkButton } from "../Layout/Buttons/DarkButton";
-import { MaterialIcons, Feather } from "@expo/vector-icons";
+import { MaterialIcons } from "@expo/vector-icons";
+import { ModalWithClose } from "../Layout/ModalwithClose";
 
 const styles = StyleSheet.create({
-  background: {
-    ...StyleSheet.absoluteFillObject,
-    backgroundColor: color("grey", 100),
-    opacity: 0.8
-  },
-  cardWrapper: {
-    padding: size(3),
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "center"
-  },
-  card: {
-    width: 512,
-    maxWidth: "100%"
-  },
   inputWrapper: {
     marginTop: size(4),
     marginBottom: size(2)
   }
 });
 
-interface ManualInputCard {
-  isVisible: boolean;
+interface ManualInputCard extends ModalWithClose {
   onVoucherCodeSubmit: (voucherCode: string) => void;
-  onExit: () => void;
 }
 
 export const ManualAddVoucherModal: FunctionComponent<ManualInputCard> = ({
@@ -54,50 +31,23 @@ export const ManualAddVoucherModal: FunctionComponent<ManualInputCard> = ({
   };
 
   return (
-    <Modal
-      visible={isVisible}
-      onRequestClose={onExit}
-      animationType="fade"
-      transparent={true}
-    >
-      <TouchableWithoutFeedback onPress={onExit}>
-        <View style={styles.background} />
-      </TouchableWithoutFeedback>
-      <View style={styles.cardWrapper}>
-        <Card style={styles.card}>
-          <TouchableOpacity
-            onPress={onExit}
-            style={{
-              position: "absolute",
-              right: size(1.5),
-              top: size(1.5),
-              padding: size(1)
-            }}
-          >
-            <Feather name="x" size={size(3)} color={color("blue", 50)} />
-          </TouchableOpacity>
-          <View style={styles.inputWrapper}>
-            <InputWithLabel
-              label="Enter voucher ID"
-              value={voucherCode}
-              onChange={({ nativeEvent: { text } }) => setVoucherCode(text)}
-              onSubmitEditing={onSubmit}
-            />
-          </View>
-          <DarkButton
-            fullWidth
-            text="Add voucher"
-            icon={
-              <MaterialIcons
-                name="add"
-                size={size(2)}
-                color={color("grey", 0)}
-              />
-            }
-            onPress={onSubmit}
-          />
-        </Card>
+    <ModalWithClose isVisible={isVisible} onExit={onExit}>
+      <View style={styles.inputWrapper}>
+        <InputWithLabel
+          label="Enter voucher ID"
+          value={voucherCode}
+          onChange={({ nativeEvent: { text } }) => setVoucherCode(text)}
+          onSubmitEditing={onSubmit}
+        />
       </View>
-    </Modal>
+      <DarkButton
+        fullWidth
+        text="Add voucher"
+        icon={
+          <MaterialIcons name="add" size={size(2)} color={color("grey", 0)} />
+        }
+        onPress={onSubmit}
+      />
+    </ModalWithClose>
   );
 };
